// src/app/core/api-v1/model/registerRequest.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterRequest { 
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    phone: string;
    role: RegisterRequest.RoleEnum;
    businessName?: string;
}
export namespace RegisterRequest {
    export const RoleEnum = {
        Camper: 'CAMPER',
        Merchant: 'MERCHANT'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}